Created a seperate component for 'WhatsNew'
